requirements.txt
cachelib==0.10.2
click==8.1.3
Flask==2.2.3
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.5
python-decouple==3.8
python-dotenv==1.0.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3
